+++ lib/main.ts
@@ -1,21 +1,16 @@
-declare const require: (path: string) => any;
-declare const __dirname: string;
-declare const process: any;
-declare const global: any;
+import { CompositeDisposable, Package } from "atom";
+import registry = require('winreg');
+import path = require('path');
+import fs = require('fs');
+import os = require("os");
 
-const _atom = global.atom;
-
-const registry = require('winreg');
-const path = require('path');
-const atom = require("atom");
-const fs = require('fs');
-const os = require("os");
-
+var myself: Package;
 const release = os.release();
 const isWin = /^win\d+$/.test(process.platform);
 const isWin8 = isWin && /^6\.[23456789]\./.test(release);
 const isWin10 = isWin && !isWin8 && /^10\./.test(release);
 const isWin10or8 = isWin10 || isWin8;
+const noop = function(){}
 
 var currentAccent: string;
 var changedAccentColor: string;
@@ -27,12 +22,12 @@ const WINDOWS_ACCENT_KEY_REG = isWin10or8 ? new registry({
   key: '\\SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Explorer\\Accent'
 }) : undefined;
 /*const setDarkColors = function() {
-  _atom.config.set("win10-ui.themeBackgroundColor", "#35373b");
-  _atom.config.set("win10-ui.themeForegroundColor", "#e8e8e8");
+  atom.config.set("win10-ui.themeBackgroundColor", "#35373b");
+  atom.config.set("win10-ui.themeForegroundColor", "#e8e8e8");
 }
 const setLightColors = function() {
-  _atom.config.set("win10-ui.themeBackgroundColor", "#e0e0e0");
-  _atom.config.set("win10-ui.themeForegroundColor", "#0a0a0a");
+  atom.config.set("win10-ui.themeBackgroundColor", "#e0e0e0");
+  atom.config.set("win10-ui.themeForegroundColor", "#0a0a0a");
 }*/
 const isDark = function(color: string) {
   const R = parseInt(color.substring(1, 3), 16);
@@ -42,7 +37,7 @@ const isDark = function(color: string) {
   return (5 * G + 2 * R + B) <= 8 * 128;
 }
 /*const setDynamicColors = function(invert?: boolean) {
-  var can = isDark(_atom.config.get('win10-ui.themeAccentColor').toHexString());
+  var can = isDark(atom.config.get('win10-ui.themeAccentColor').toHexString());
   if(invert)
     can = !can;
   if(can)
@@ -50,17 +45,19 @@ const isDark = function(color: string) {
   else
     setLightColors();
 }*/
+var debug: typeof console.log;
+var trace: typeof console.trace;
 const writeConfig = function(color?: any) {
-  const accentColor = (color || _atom.config.get('win10-ui.themeAccentColor')).toHexString();
-  //const backgroundColor = _atom.config.get('win10-ui.themeBackgroundColor').toHexString();
+  const accentColor = (color || atom.config.get('win10-ui.themeAccentColor')).toHexString();
+  //const backgroundColor = atom.config.get('win10-ui.themeBackgroundColor').toHexString();
   const accentColorDark = isDark(accentColor);
-  console.log("writeConfig", accentColor, accentColorDark);
-  const config = `@font-size: ${_atom.config.get('win10-ui.fontSize')}px;\n` +
+  debug("writeConfig", accentColor, accentColorDark);
+  const config = `@font-size: ${atom.config.get('win10-ui.fontSize')}px;\n` +
     `@accent-color: ${accentColor};\n` +
     `@accent-color-is-dark: ${accentColorDark ? 1 : 0};\n` +
     /*`@base-background-color: ${backgroundColor};\n` +
     `@background-color-dark: ${isDark(backgroundColor) ? 100 : 0}%;\n\n` +
-    `@text-color: ${_atom.config.get('win10-ui.themeForegroundColor').toHexString()};\n` +*/
+    `@text-color: ${atom.config.get('win10-ui.themeForegroundColor').toHexString()};\n` +*/
     `@text-color-accent:  ${accentColorDark ? "lighten(@text-color, 40%)" : "darken(@text-color, 40%)"};\n` +
     `@text-color-accent-contrast:  ${accentColorDark ? "white" : "black"};`;
   fs.readFile(path.resolve(stylesPath, "ui-variables.less.input"), "utf-8", function(err, inputVariables) {
@@ -71,14 +68,12 @@ const writeConfig = function(color?: any) {
         if(err)
           throw err;
         else
-          for (const theme of _atom.themes.getActiveThemes()) {
-            theme.reloadStylesheets();
-          }
+          atom.packages.getActivePackage("win10-ui")['reloadStylesheets']();
       });
   });
 
 }
-var updateInterval: number;
+var updateInterval: NodeJS.Timer;
 const updateAccent = function(writeAlways?: boolean) {
   return WINDOWS_ACCENT_KEY_REG.get(ACCENT_VALUE, function(error, item) {
     if (error) {
@@ -96,11 +91,11 @@ const updateAccent = function(writeAlways?: boolean) {
       return;
     }
     currentAccent = color;
-    console.log("read accent color", color);
-    if(_atom.config.get('win10-ui.themeAccentColor').toHexString() !== color) {
+    debug("read accent color", color);
+    if(atom.config.get('win10-ui.themeAccentColor').toHexString() !== color) {
       changedAccentColor = color;
-      console.log("setting accent color", color);
-      _atom.config.set('win10-ui.themeAccentColor', color);
+      debug("setting accent color", color);
+      atom.config.set('win10-ui.themeAccentColor', color);
     } else if (writeAlways)
       writeConfig();
   });
@@ -113,7 +108,7 @@ const reset = function() {
 var presetChanged: boolean;
 /*const updatePresetAndWriteConfig = function() {
   if(!changedAccentColor)
-    _atom.config.set(`win10-ui.preset`, "Custom");
+    atom.config.set(`win10-ui.preset`, "Custom");
   writeConfig();
 }*/
 
@@ -160,11 +155,26 @@ const win10 = {
       minimum: 8,
       maximum: 20,
       default: 12
+    },
+    debug: {
+      order: 7,
+      description: 'Output to the console',
+      type: 'boolean',
+      default: false
     }
   },
   activate: function(state) {
-    compositeDisposable = new atom.CompositeDisposable;
-    /*compositeDisposable.add(_atom.config.onDidChange(`win10-ui.preset`, function(preset) {
+    if (atom.config.get("win10-ui.debug")) {
+      debug = console.log.bind(console);
+      trace = console.log.bind(console);
+    } else {
+      debug = noop;
+      trace = noop;
+    }
+
+    myself = atom.packages.getActivePackage("win10-ui");
+    compositeDisposable = new CompositeDisposable;
+    /*compositeDisposable.add(atom.config.onDidChange(`win10-ui.preset`, function(preset) {
       changedAccentColor = true;
       switch(preset.newValue) {
         case "Dark":
@@ -181,36 +191,45 @@ const win10 = {
         changedAccentColor = false;
       }, 250);
     }));*/
-    compositeDisposable.add(_atom.config.onDidChange(`win10-ui.fontSize`, writeConfig));
-    //compositeDisposable.add(_atom.config.onDidChange(`win10-ui.themeForegroundColor`, updatePresetAndWriteConfig));
-    //compositeDisposable.add(_atom.config.onDidChange(`win10-ui.themeBackgroundColor`, updatePresetAndWriteConfig));
-    compositeDisposable.add(_atom.config.onDidChange(`win10-ui.useSystemAccentColor`, function(data) {
+    compositeDisposable.add(atom.config.onDidChange(`win10-ui.debug`, function(data) {
+      if (data.newValue) {
+        debug = console.log.bind(console);
+        trace = console.trace.bind(console);
+      } else {
+        debug = noop;
+        trace = noop;
+      }
+    }));
+    compositeDisposable.add(atom.config.onDidChange(`win10-ui.fontSize`, writeConfig));
+    //compositeDisposable.add(atom.config.onDidChange(`win10-ui.themeForegroundColor`, updatePresetAndWriteConfig));
+    //compositeDisposable.add(atom.config.onDidChange(`win10-ui.themeBackgroundColor`, updatePresetAndWriteConfig));
+    compositeDisposable.add(atom.config.onDidChange(`win10-ui.useSystemAccentColor`, function(data) {
       try{clearInterval(updateInterval)}catch(e){}
       if (isWin10or8 && data.newValue) {
         updateInterval = setInterval(updateAccent, 1500);
-        console.log("Started updateAccent interval");
+        debug("Started updateAccent interval");
         currentAccent = undefined;
         updateAccent();
-      } else
-        console.log("Stopped updateAccent interval");
+      } else {
+        debug("Stopped updateAccent interval");
+        trace();
+      }
     }));
-    compositeDisposable.add(_atom.config.onDidChange(`win10-ui.themeAccentColor`, function(data) {
-      if(!changedAccentColor) {
-        console.log("use system accent disabled because of user interaction");
-        _atom.config.set(`win10-ui.useSystemAccentColor`, false);
-      } else if(changedAccentColor) {
-        if (data.newValue.toHexString() !== changedAccentColor)
-          return;
-        console.log("accent color changed dynamically");
+    compositeDisposable.add(atom.config.onDidChange(`win10-ui.themeAccentColor`, function(data) {
+      if(data.newValue.toHexString() !== changedAccentColor) {
+        debug("use system accent disabled because of user interaction");
+        atom.config.set(`win10-ui.useSystemAccentColor`, false);
+      } else {
+        debug("accent color changed dynamically");
         changedAccentColor = undefined;
       }
-      console.log("accent color changed", data.newValue);
+      debug("accent color changed", data.newValue);
       writeConfig(data.newValue);
     }));
-    if (isWin10or8 && _atom.config.get("win10-ui.useSystemAccentColor")) {
+    if (isWin10or8 && atom.config.get("win10-ui.useSystemAccentColor")) {
       try{clearInterval(updateInterval)}catch(e){}
       updateInterval = setInterval(updateAccent, 1500);
-      console.log("Started updateAccent interval");
+      debug("Started updateAccent interval");
       currentAccent = undefined;
       updateAccent(true);
     } else